chore(flutter): split unit and integration tests into separate actions
